card refactors
