+++ explorer-nextjs/src/components/landingPageComponents/NoiseCard.tsx
@@ -1,6 +1,6 @@
 import type { IPacketsAndStakingData } from "@/app/api";
 import { HARBOURMASTER_API_MIXNODES_STATS } from "@/app/api/urls";
-import { formatBigNum } from "@/app/utils/formatBigNumbers";
+import { formatBigNum } from "@/utils/formatBigNumbers";
 import { Box, Stack, Typography } from "@mui/material";
 import ExplorerCard from "../cards/ExplorerCard";
 import { LineChart } from "../lineChart";
