+++ explorer-nextjs/src/utils/formatBigNumbers.ts
